Struct nonempty_vec::NonemptyVec

pub struct NonemptyVec<T, H: HeadLocation = HeadFirst> { /* fields omitted */ }

A Vec that always has at least one element

Implementations

impl<T, H: HeadLocation> NonemptyVec<T, H>

pub fn new(head: T) -> Self

Creates a new NonemptyVec with one element

pub fn head(&self) -> &T

The head is simply the element guaranteed to exist. It is not necessarily interpreted as the first item

pub fn head_mut(&mut self) -> &mut T

Like head but mutable

pub fn into_head(self) -> T

Consumes self and returns the head

pub fn tail(&self) -> &[T]

The tail is simply all elements that are not guaranteed to exist

pub fn tail_mut(&mut self) -> &mut Vec<T>

Like tail but mutable

pub fn into_tail(self) -> Vec<T>

Consumes self and returns the tail

pub fn into_head_tail(self) -> (T, Vec<T>)

Consumes self and returns the head and the tail

pub fn len(&self) -> NonZeroUsize

Returns the length of self

Edge Case

The length can be usize::MAX + 1 - in that case, only usize::MAX is returned

To circumvent this, you can work with self.tail().len(), which is always one less than self.len()

pub fn push(&mut self, t: T)

Pushes an element to the end of self

pub fn pop(&mut self) -> Option<T>

Removes the last element from self (unless self has only one item left), returns it.

pub fn insert(&mut self, index: usize, element: T)

Inserts an element at the specified index

pub fn remove(&mut self, index: usize) -> Option<T>

Removes an element from the specified index; Returns None if there is only one element left

pub fn into_vec(self) -> Vec<T>

Creates a Vec from all its elements

pub fn first(&self) -> &T

The first element

pub fn first_mut(&mut self) -> &mut T

The first element, mutable

pub fn into_first(self) -> T

Returns the current first element, consuming self

pub fn last(&self) -> &T

The last element

pub fn last_mut(&mut self) -> &mut T

The last element, mutable

pub fn into_last(self) -> T

Returns the current last element, consuming self