Trait nonempty_vec::HeadLocation

pub trait HeadLocation: Sealed {
    const HEAD_FIRST: bool;
}

Specifies the location the head item has in relation to the rest

Associated Constants

const HEAD_FIRST: bool

This constant can be used to write code generic over HeadLocation without explicit casing

Loading content...